+++ go/plugin/caddy2/dal/models/admin_api_audit_log_models.go
@@ -3,44 +3,44 @@ package models
 import (
 	"context"
 	"errors"
-	"github.com/inverse-inc/packetfence/go/caddy/pfpki/sql"
-	"github.com/jinzhu/gorm"
 	"strings"
 	"time"
+
+	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
+	"github.com/jinzhu/gorm"
 )
 
-type AuthLog struct {
-	ID          int64      `json:"id,omitempty" gorm:"primary_key"`
-	ProcessName string     `json:"process_name,omitempty"`
-	Mac         string     `json:"mac,omitempty"`
-	Pid         string     `json:"pid,omitempty" gorm:"default:'default'"`
-	Status      string     `json:"status,omitempty" gorm:"default:'incomplete'"`
-	AttemptedAt *time.Time `json:"attempted_at"`
-	CompletedAt *time.Time `json:"completed_at"`
-	Source      string     `json:"source,omitempty"`
-	Profile     string     `json:"profile,omitempty"`
+type AdminApiAuditLog struct {
+	ID        int64      `json:"id,omitempty" gorm:"primary_key"`
+	CreatedAt *time.Time `json:"created_at,omitempty"`
+	UserName  string     `json:"user_name,omitempty"`
+	Url       string     `json:"url,omitempty"`
+	Action    string     `json:"action,omitempty"`
+	ObjectId  string     `json:"object_id,omitempty"`
+	Method    string     `json:"method,omitempty"`
+	Request   string     `json:"request,omitempty"`
+	Status    int16      `json:"status,omitempty"`
 
 	DB  *gorm.DB         `json:"-" gorm:"-"`
 	Ctx *context.Context `json:"-" gorm:"-"`
 }
 
-func (a AuthLog) TableName() string {
-	return "auth_log"
+func (a AdminApiAuditLog) TableName() string {
+	return "admin_api_audit_log"
 }
 
-// AuthLog
-func NewAuthLogModel(db *gorm.DB, ctx *context.Context) *AuthLog {
-	ret := &AuthLog{}
+func NewAdminApiAuditLogModel(db *gorm.DB, ctx *context.Context) *AdminApiAuditLog {
+	ret := &AdminApiAuditLog{}
 	ret.DB = db
 	ret.Ctx = ctx
 	return ret
 }
 
-func (a AuthLog) Paginated(vars sql.Vars) (DBRes, error) {
+func (a AdminApiAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 	var res = DBRes{}
 	var count int
 
-	a.DB.Model(&AuthLog{}).Count(&count)
+	a.DB.Model(&AdminApiAuditLog{}).Count(&count)
 	res.Total = &count
 	res.PrevCursor = &vars.Cursor
 	nextCursor := vars.Cursor + vars.Limit
@@ -51,7 +51,7 @@ func (a AuthLog) Paginated(vars sql.Vars) (DBRes, error) {
 		if err != nil {
 			return DBRes{}, err
 		}
-		var items []AuthLog
+		var items []AdminApiAuditLog
 		db := a.DB.Select(sqls.Select).Order(sqls.Order).Offset(sqls.Offset).Limit(sqls.Limit).Find(&items)
 		if db.Error != nil {
 			return DBRes{}, db.Error
@@ -61,7 +61,7 @@ func (a AuthLog) Paginated(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a AuthLog) Search(vars sql.Vars) (DBRes, error) {
+func (a AdminApiAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	res := DBRes{}
 	sqls, err := vars.Sql(a)
 	if err != nil {
@@ -69,8 +69,8 @@ func (a AuthLog) Search(vars sql.Vars) (DBRes, error) {
 	}
 
 	var count int
-	var items []AuthLog
-	a.DB.Model(&AuthLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
+	var items []AdminApiAuditLog
+	a.DB.Model(&AdminApiAuditLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
 
 	if count == 0 {
 		return res, errors.New("entries not found")
@@ -91,9 +91,9 @@ func (a AuthLog) Search(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a AuthLog) GetByID(id string) (DBRes, error) {
+func (a AdminApiAuditLog) GetByID(id string) (DBRes, error) {
 	res := DBRes{}
-	var item AuthLog
+	var item AdminApiAuditLog
 
 	allFields := strings.Join(sql.SqlFields(a)[:], ",")
 	db := a.DB.Select(allFields).Where("`id` = ?", id).First(&item)
@@ -103,9 +103,9 @@ func (a AuthLog) GetByID(id string) (DBRes, error) {
 	return res, db.Error
 }
 
-func (a AuthLog) Delete(id string) (DBRes, error) {
+func (a AdminApiAuditLog) Delete(id string) (DBRes, error) {
 	res := DBRes{}
-	var item AuthLog
+	var item AdminApiAuditLog
 	db := a.DB.Where("`id` = ?", id).Find(&item)
 	if db.Error != nil {
 		return res, db.Error
@@ -114,11 +114,11 @@ func (a AuthLog) Delete(id string) (DBRes, error) {
 	return res, err
 }
 
-func (a AuthLog) Update() (DBRes, error) {
-	var item AuthLog
+func (a AdminApiAuditLog) Update() (DBRes, error) {
+	var item AdminApiAuditLog
 	res := DBRes{}
 
-	err := a.DB.Model(&AuthLog{}).Where("id = ?", a.ID).Updates(a).Error
+	err := a.DB.Model(&AdminApiAuditLog{}).Where("id = ?", a.ID).Updates(a).Error
 
 	if err != nil {
 		return DBRes{}, err
