+++ go/plugin/caddy2/dal/models/dns_audit_log.go
@@ -3,43 +3,43 @@ package models
 import (
 	"context"
 	"errors"
-	"github.com/inverse-inc/packetfence/go/caddy/pfpki/sql"
-	"github.com/jinzhu/gorm"
 	"strings"
 	"time"
+
+	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
+	"github.com/jinzhu/gorm"
 )
 
-type AdminApiAuditLog struct {
+type DnsAuditLog struct {
 	ID        int64      `json:"id,omitempty" gorm:"primary_key"`
 	CreatedAt *time.Time `json:"created_at,omitempty"`
-	UserName  string     `json:"user_name,omitempty"`
-	Url       string     `json:"url,omitempty"`
-	Action    string     `json:"action,omitempty"`
-	ObjectId  string     `json:"object_id,omitempty"`
-	Method    string     `json:"method,omitempty"`
-	Request   string     `json:"request,omitempty"`
-	Status    int16      `json:"status,omitempty"`
+	IP        string     `json:"ip,omitempty"`
+	Mac       string     `json:"mac,omitempty"`
+	QName     string     `json:"qname,omitempty" gorm:"column:qname"`
+	QType     string     `json:"qtype,omitempty" gorm:"column:qtype"`
+	Scope     string     `json:"scope,omitempty"`
+	Answer    string     `json:"answer,omitempty"`
 
 	DB  *gorm.DB         `json:"-" gorm:"-"`
 	Ctx *context.Context `json:"-" gorm:"-"`
 }
 
-func (a AdminApiAuditLog) TableName() string {
-	return "admin_api_audit_log"
+func (d DnsAuditLog) TableName() string {
+	return "dns_audit_log"
 }
 
-func NewAdminApiAuditLogModel(db *gorm.DB, ctx *context.Context) *AdminApiAuditLog {
-	ret := &AdminApiAuditLog{}
+// DnsAuditLog
+func NewDnsAuditLogModel(db *gorm.DB, ctx *context.Context) *DnsAuditLog {
+	ret := &DnsAuditLog{}
 	ret.DB = db
 	ret.Ctx = ctx
 	return ret
 }
-
-func (a AdminApiAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
+func (a DnsAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 	var res = DBRes{}
 	var count int
 
-	a.DB.Model(&AdminApiAuditLog{}).Count(&count)
+	a.DB.Model(&DnsAuditLog{}).Count(&count)
 	res.Total = &count
 	res.PrevCursor = &vars.Cursor
 	nextCursor := vars.Cursor + vars.Limit
@@ -50,7 +50,7 @@ func (a AdminApiAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 		if err != nil {
 			return DBRes{}, err
 		}
-		var items []AdminApiAuditLog
+		var items []DnsAuditLog
 		db := a.DB.Select(sqls.Select).Order(sqls.Order).Offset(sqls.Offset).Limit(sqls.Limit).Find(&items)
 		if db.Error != nil {
 			return DBRes{}, db.Error
@@ -60,7 +60,7 @@ func (a AdminApiAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a AdminApiAuditLog) Search(vars sql.Vars) (DBRes, error) {
+func (a DnsAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	res := DBRes{}
 	sqls, err := vars.Sql(a)
 	if err != nil {
@@ -68,8 +68,8 @@ func (a AdminApiAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	}
 
 	var count int
-	var items []AdminApiAuditLog
-	a.DB.Model(&AdminApiAuditLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
+	var items []DnsAuditLog
+	a.DB.Model(&DnsAuditLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
 
 	if count == 0 {
 		return res, errors.New("entries not found")
@@ -90,9 +90,9 @@ func (a AdminApiAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a AdminApiAuditLog) GetByID(id string) (DBRes, error) {
+func (a DnsAuditLog) GetByID(id string) (DBRes, error) {
 	res := DBRes{}
-	var item AdminApiAuditLog
+	var item DnsAuditLog
 
 	allFields := strings.Join(sql.SqlFields(a)[:], ",")
 	db := a.DB.Select(allFields).Where("`id` = ?", id).First(&item)
@@ -102,9 +102,9 @@ func (a AdminApiAuditLog) GetByID(id string) (DBRes, error) {
 	return res, db.Error
 }
 
-func (a AdminApiAuditLog) Delete(id string) (DBRes, error) {
+func (a DnsAuditLog) Delete(id string) (DBRes, error) {
 	res := DBRes{}
-	var item AdminApiAuditLog
+	var item DnsAuditLog
 	db := a.DB.Where("`id` = ?", id).Find(&item)
 	if db.Error != nil {
 		return res, db.Error
@@ -113,11 +113,11 @@ func (a AdminApiAuditLog) Delete(id string) (DBRes, error) {
 	return res, err
 }
 
-func (a AdminApiAuditLog) Update() (DBRes, error) {
-	var item AdminApiAuditLog
+func (a DnsAuditLog) Update() (DBRes, error) {
+	var item DnsAuditLog
 	res := DBRes{}
 
-	err := a.DB.Model(&AdminApiAuditLog{}).Where("id = ?", a.ID).Updates(a).Error
+	err := a.DB.Model(&DnsAuditLog{}).Where("id = ?", a.ID).Updates(a).Error
 
 	if err != nil {
 		return DBRes{}, err
