+++ go/plugin/caddy2/dal/models/radacct_log.go
@@ -3,42 +3,45 @@ package models
 import (
 	"context"
 	"errors"
-	"github.com/inverse-inc/packetfence/go/caddy/pfpki/sql"
-	"github.com/jinzhu/gorm"
 	"strings"
 	"time"
+
+	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
+	"github.com/jinzhu/gorm"
 )
 
-type DnsAuditLog struct {
-	ID        int64      `json:"id,omitempty" gorm:"primary_key"`
-	CreatedAt *time.Time `json:"created_at,omitempty"`
-	IP        string     `json:"ip,omitempty"`
-	Mac       string     `json:"mac,omitempty"`
-	QName     string     `json:"qname,omitempty" gorm:"column:qname"`
-	QType     string     `json:"qtype,omitempty" gorm:"column:qtype"`
-	Scope     string     `json:"scope,omitempty"`
-	Answer    string     `json:"answer,omitempty"`
+type RadacctLog struct {
+	ID               int64      `json:"id,omitempty" gorm:"primary_key"`
+	AcctSessionID    string     `json:"acctsessionid,omitempty" gorm:"column:acctsessionid;default:''"`
+	UserName         string     `json:"username,omitempty" gorm:"column:username;default:''"`
+	NasIPAddress     string     `json:"nasipaddress,omitempty" gorm:"column:nasipaddress;default:''"`
+	AcctStatusType   string     `json:"acctstatustype,omitempty" gorm:"column:acctstatustype;default:''"`
+	Timestamp        *time.Time `json:"timestamp,omitempty"`
+	AcctInputOctets  int64      `json:"acctinputoctets,omitempty" gorm:"column:acctinputoctets"`
+	AcctOutputOctets int64      `json:"acctoutputoctets,omitempty" gorm:"column:acctoutputoctets"`
+	AcctSessionTime  int        `json:"acctsessiontime,omitempty" gorm:"column:acctsessiontime"`
+	AcctUniqueID     string     `json:"acctuniqueid,omitempty" gorm:"column:acctuniqueid;default:''"`
 
 	DB  *gorm.DB         `json:"-" gorm:"-"`
 	Ctx *context.Context `json:"-" gorm:"-"`
 }
 
-func (d DnsAuditLog) TableName() string {
-	return "dns_audit_log"
+func (r RadacctLog) TableName() string {
+	return "radacct_log"
 }
 
-// DnsAuditLog
-func NewDnsAuditLogModel(db *gorm.DB, ctx *context.Context) *DnsAuditLog {
-	ret := &DnsAuditLog{}
+// RadacctLog
+func NewRadacctLogModel(db *gorm.DB, ctx *context.Context) *RadacctLog {
+	ret := &RadacctLog{}
 	ret.DB = db
 	ret.Ctx = ctx
 	return ret
 }
-func (a DnsAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
+func (a RadacctLog) Paginated(vars sql.Vars) (DBRes, error) {
 	var res = DBRes{}
 	var count int
 
-	a.DB.Model(&DnsAuditLog{}).Count(&count)
+	a.DB.Model(&RadacctLog{}).Count(&count)
 	res.Total = &count
 	res.PrevCursor = &vars.Cursor
 	nextCursor := vars.Cursor + vars.Limit
@@ -49,7 +52,7 @@ func (a DnsAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 		if err != nil {
 			return DBRes{}, err
 		}
-		var items []DnsAuditLog
+		var items []RadacctLog
 		db := a.DB.Select(sqls.Select).Order(sqls.Order).Offset(sqls.Offset).Limit(sqls.Limit).Find(&items)
 		if db.Error != nil {
 			return DBRes{}, db.Error
@@ -59,7 +62,7 @@ func (a DnsAuditLog) Paginated(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a DnsAuditLog) Search(vars sql.Vars) (DBRes, error) {
+func (a RadacctLog) Search(vars sql.Vars) (DBRes, error) {
 	res := DBRes{}
 	sqls, err := vars.Sql(a)
 	if err != nil {
@@ -67,8 +70,8 @@ func (a DnsAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	}
 
 	var count int
-	var items []DnsAuditLog
-	a.DB.Model(&DnsAuditLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
+	var items []RadacctLog
+	a.DB.Model(&RadacctLog{}).Where(sqls.Where.Query, sqls.Where.Values...).Count(&count)
 
 	if count == 0 {
 		return res, errors.New("entries not found")
@@ -89,9 +92,9 @@ func (a DnsAuditLog) Search(vars sql.Vars) (DBRes, error) {
 	return res, nil
 }
 
-func (a DnsAuditLog) GetByID(id string) (DBRes, error) {
+func (a RadacctLog) GetByID(id string) (DBRes, error) {
 	res := DBRes{}
-	var item DnsAuditLog
+	var item RadacctLog
 
 	allFields := strings.Join(sql.SqlFields(a)[:], ",")
 	db := a.DB.Select(allFields).Where("`id` = ?", id).First(&item)
@@ -101,9 +104,9 @@ func (a DnsAuditLog) GetByID(id string) (DBRes, error) {
 	return res, db.Error
 }
 
-func (a DnsAuditLog) Delete(id string) (DBRes, error) {
+func (a RadacctLog) Delete(id string) (DBRes, error) {
 	res := DBRes{}
-	var item DnsAuditLog
+	var item RadacctLog
 	db := a.DB.Where("`id` = ?", id).Find(&item)
 	if db.Error != nil {
 		return res, db.Error
@@ -112,11 +115,11 @@ func (a DnsAuditLog) Delete(id string) (DBRes, error) {
 	return res, err
 }
 
-func (a DnsAuditLog) Update() (DBRes, error) {
-	var item DnsAuditLog
+func (a RadacctLog) Update() (DBRes, error) {
+	var item RadacctLog
 	res := DBRes{}
 
-	err := a.DB.Model(&DnsAuditLog{}).Where("id = ?", a.ID).Updates(a).Error
+	err := a.DB.Model(&RadacctLog{}).Where("id = ?", a.ID).Updates(a).Error
 
 	if err != nil {
 		return DBRes{}, err
